adds path for passing a query to copy from sql to pg instead of table name

+++ ris/copy_table_from_sql_to_pg.py
@@ -5,6 +5,24 @@ import datetime
 import os
 
 
+def get_table_fromsql_query(sql, qry):
+    # assumes full query including schema etc passed
+    q = sql.query("""select top 1 * from ({q}) as q""".format(q=qry))
+    path = os.path.join(os.getcwd(), 'data.csv')
+
+    cmd = """sqlcmd -S {serv} -d {db} -U {usr} -P {pas} 
+        -Q "set nocount on; select * from ({q}) as q" 
+        -o {p} -h-1 -s "|" -W """.format(db=sql.params['DATABASE'],
+                                         serv=sql.params['SERVER'],
+                                         pas=sql.params['PWD'],
+                                         usr=sql.params['UID'],
+                                         q=qry,
+                                         p=path)
+    os.system(cmd.replace('\n', ' '))
+    clean_output(path)
+    return q._replace(data=None)
+
+
 def get_table_from_sql(sql, sql_schema, sql_table):
     # TODO: add option to pass in query instead of full table
     q = sql.query("""select top 1 * from {db}.{sch}.{tbl}""".format(db=sql.params['DATABASE'],
@@ -39,11 +57,20 @@ def clean_output(path):
 
 def add_table_to_pgsql(pg, pg_schema, pg_table, table_data, archive=False, permission_default=False):
     if archive:
-        pg.query("drop table if exists archive.{}".format(pg_table))
-        pg.query("create table archive.{t} as select * from {s}.{t}".format(t=pg_table, s=pg_schema))
-        pg.query("alter table archive.{t} add archived_dte timestamp".format(
-            t=pg_table, s=pg_schema))
-        pg.query("update archive.{t} set archived_dte=now()".format(t=pg_table, s=pg_schema))
+        exists = pg.query("""SELECT EXISTS (
+                            SELECT 1
+                            FROM   information_schema.tables
+                            WHERE  table_schema = '{s}'
+                            AND    table_name = '{t}'
+                           );""".format(t=pg_table, s=pg_schema))
+        if exists.data[0][0]:
+            pg.query("drop table if exists archive.{}".format(pg_table))
+            pg.query("create table archive.{t} as select * from {s}.{t}".format(t=pg_table, s=pg_schema))
+            pg.query("alter table archive.{t} add archived_dte timestamp".format(
+                t=pg_table, s=pg_schema))
+            pg.query("update archive.{t} set archived_dte=now()".format(t=pg_table, s=pg_schema))
+        else:
+            print 'No existing table to archive'
 
     pg.query("drop table if exists {s}.{t}".format(t=pg_table, s=pg_schema))
     pg.query(build_table(pg_schema, pg_table, table_data))
@@ -81,7 +108,9 @@ def build_table(pg_schema, pg_table, table_data):
 
 
 def connection_ui(sql=None, pg=None):
-    if not sql: # passed in sql connection
+    q = raw_input('Get data from full table (T) or query (Q)\n').upper()
+
+    if not sql:  # passed in sql connection
         org_server = raw_input('\nOrigin SQL server name:\n')
         org_db = raw_input('\nOrigin SQL databse name:\n')
         org_user = raw_input('\nOrigin db login user name:\n')
@@ -89,8 +118,9 @@ def connection_ui(sql=None, pg=None):
     else:
         org_server = sql.params['SERVER']
         org_db = sql.params['DATABASE']
-    org_schema = raw_input('\nOrigin SQL schema name:\n')
-    org_table = raw_input('\nOrigin SQL table name:\n')
+    if q == 'T':
+        org_schema = raw_input('\nOrigin SQL schema name:\n')
+        org_table = raw_input('\nOrigin SQL table name:\n')
 
     if not pg:  # passed in sql connection
         dest_server = raw_input('\nDestination PG server name:\n')
@@ -104,12 +134,17 @@ def connection_ui(sql=None, pg=None):
 
     dest_schema = raw_input('\nDestination PG schema name:\n')
 
-    if raw_input('\nRename table? (Y/N)\n').lower() == 'y':
-        dest_table = raw_input('Destination table name:\n')
+    if q == 'T':
+        if raw_input('\nRename table? (Y/N)\n').lower() == 'y':
+            dest_table = raw_input('Destination table name:\n')
+        else:
+            dest_table = org_table
+        table = get_table_from_sql(sql, org_schema, org_table)
     else:
-        dest_table = org_table
+        dest_table = raw_input('\nNew table name:\n').lower()
+        qry = raw_input('QUERY =:\n')
+        table = get_table_fromsql_query(sql, qry)
 
-    table = get_table_from_sql(sql, org_schema, org_table)
     add_table_to_pgsql(pg, dest_schema, dest_table, table, True, True)
